move dispatch parameters to a separate class (this way we can set the reference to null and avoid accessing params when they're invalid, i.e. StatusFileSet == false)
