+++ VSRAD.Package/VSRAD.Package.csproj
@@ -214,6 +214,7 @@
     <Compile Include="Server\ActionRunner.cs" />
     <Compile Include="Server\ActionRunResult.cs" />
     <Compile Include="Server\BreakStateData.cs" />
+    <Compile Include="Server\BreakStateDispatchParameters.cs" />
     <Compile Include="Server\DebugRunResult.cs" />
     <Compile Include="ToolWindows\EvaluateSelectedControl.xaml.cs">
       <DependentUpon>EvaluateSelectedControl.xaml</DependentUpon>
